refactor: :recycle: refactoring all charts to vue-3, removin vue-chart dependency
