+++ package.json
@@ -33,7 +33,7 @@
     "<rootDir>/.jest/register-context.js"
   ],
   "dependencies": {
-    "@vue/compat": "^3.2.27",
+    "@vue/compat": "^3.2.29",
     "chart.js": "2.9.4",
     "core-js": "3.20.2",
     "html2canvas": "^1.3.3",
@@ -42,9 +42,7 @@
     "maska": "^1.5.0",
     "tailwindcss": "^1.9.6",
     "v-mask": "^2.3.0",
-    "vue": "^3.2.27",
-    "vue-chart-3": "^3.0.7",
-    "vue-chartjs": "^3.5.1",
+    "vue": "^3.2.29",
     "vue2-datepicker": "3.10.4"
   },
   "devDependencies": {
@@ -63,7 +61,7 @@
     "@vue/cli-plugin-unit-jest": "~4.5.15",
     "@vue/cli-plugin-vuex": "~4.5.15",
     "@vue/cli-service": "~4.5.15",
-    "@vue/compiler-sfc": "^3.2.27",
+    "@vue/compiler-sfc": "^3.2.29",
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/test-utils": "^2.0.0-rc.18",
     "@vue/vue3-jest": "^27.0.0-alpha.4",
@@ -83,6 +81,7 @@
     "ngrok": "^4.2.2",
     "node-sass": "4.14.1",
     "plop": "3.0.5",
+    "postcss": "^8.4.5",
     "prettier": "2.5.1",
     "resolve-url-loader": "^4.0.0",
     "sass-loader": "^8.0.2",
