+++ src/main.js
@@ -1,7 +1,14 @@
-import { createApp } from 'vue';
+import Vue, { createApp } from 'vue';
 import App from './App.vue';
 
 import '@scss/core/index.scss';
+
+Vue.configureCompat({
+  COMPONENT_V_MODEL: false,
+  CUSTOM_DIR: true,
+  RENDER_FUNCTION: false,
+  WATCH_ARRAY: false,
+});
 
 const app = createApp(App);
